+++ C++/src/Implementations/Player.h
@@ -19,36 +19,36 @@ public:
     Player(const std::string& playerName,
         const std::string& playerType);
 
-	void SetHP(int newHP) override;
-    void LoseHP(int damage) override;
+	void SetHP(uint8_t newHP) override;
+    void LoseHP(uint8_t damage) override;
     void RegainHP() override;
     void ResetHP() override;
-    int GetHP() const override;
+    int8_t GetHP() const override;
 
     bool IsAlive() const override;
     bool GetType() const override;
     std::string GetName() const override;
 
-    int GetFreeSlots() const override;
-    int GetSaws() const override;
-    int GetBeers() const override;
-    int GetMagnifiers() const override;
-    int GetHandCuffs() const override;
-    int GetInverters() const override;
-    int GetCellPhones() const override;
+    uint8_t GetFreeSlots() const override;
+    uint8_t GetSaws() const override;
+    uint8_t GetBeers() const override;
+    uint8_t GetMagnifiers() const override;
+    uint8_t GetHandCuffs() const override;
+    uint8_t GetInverters() const override;
+    uint8_t GetCellPhones() const override;
     const std::vector<uint8_t>& GetInventory() const override;
 
-    void SetFreeSlots(int newFree) override;
-	void SetSaws(int newSaws) override;
-	void SetBeers(int newBeers) override;
-	void SetMagnifiers(int newMagnifiers) override;
-	void SetHandCuffs(int newHandCuffs) override;
-	void SetInverters(int newInverters) override;
-	void SetCellPhones(int newCellPhones) override;
+    void SetFreeSlots(uint8_t newFree) override;
+	void SetSaws(uint8_t newSaws) override;
+	void SetBeers(uint8_t newBeers) override;
+	void SetMagnifiers(uint8_t newMagnifiers) override;
+	void SetHandCuffs(uint8_t newHandCuffs) override;
+	void SetInverters(uint8_t newInverters) override;
+	void SetCellPhones(uint8_t newCellPhones) override;
 	void SetInventory(const std::vector<uint8_t>& newInventory) override;
 
     void GetRandomItem() override;
     void GetNumberOfItems() override;
     void ResetInventory() override;
-    void UseItem(int itemType) override;
+    void UseItem(uint8_t itemType) override;
 };
